safari-600.7-branch/LayoutTests/fast/selectors/querySelector-simple-adjacent-backtracking.html

<!doctype html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<div style="display:none">
  <ul>
    <li></li>
    <div></div>
    <li></li>
    <blockquote></blockquote>
    <li></li>
    <li id=target1>Target 1</li>
    <li id=target2>Target 2</li>
  </ul>
</div>
</body>
<script>
description('Test backtracking with a single direct adjacent relation.');

shouldBe('document.querySelectorAll("blockquote+li~li").length', '2');
shouldBeEqualToString('document.querySelectorAll("blockquote+li~li")[0].id', 'target1');
shouldBeEqualToString('document.querySelectorAll("blockquote+li~li")[1].id', 'target2');
</script>
<script src="../../resources/js-test-post.js"></script>
</html>
