Safari-609.1.11.1/LayoutTests/fast/selectors/querySelector-simple-adjacent-backtracking-expected.txt

Test backtracking with a single direct adjacent relation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll("blockquote+li~li").length is 2
PASS document.querySelectorAll("blockquote+li~li")[0].id is "target1"
PASS document.querySelectorAll("blockquote+li~li")[1].id is "target2"
PASS successfullyParsed is true

TEST COMPLETE
